backend/requirements.txt
# --- Web Framework & Server ---
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
# --- Data Validation ---
pydantic==2.5.3
PyYAML==6.0.1
# --- Ansible Integration ---
ansible-core==2.16.2
ansible-runner==2.3.4
# --- Networking & SSH ---
# (Using a modern version to support Ed25519 and avoid the q-bit error)
paramiko==3.4.0
cryptography==42.0.2
# --- Utilities ---
requests==2.31.0
psutil==5.9.8
watchfiles==0.21.0
pymysql==1.1.0

